Refactored Questionnaire
